refactor(s3): unify all WAL requests
1. unify all WAL requests
