+++ core/src/main/scala/kafka/log/s3/objects/CommitWALObjectRequest.java
@@ -22,18 +22,41 @@ import java.util.LinkedList;
 import java.util.List;
 
 public class CommitWALObjectRequest {
+
+    /**
+     * The object id of the WAL object.
+     */
     private long objectId;
+
+    /**
+     * The order id of the WAL object.
+     * <p>
+     * When the WAL object is generated by compacting, the order id is the first compacted object's order id.
+     */
+    private long orderId;
+
+    /**
+     * The real size of the WAL object in Storage.
+     */
     private long objectSize;
     /**
-     * The stream ranges of the compacted object.
+     * The stream ranges of the WAL object.
+     * <p>
+     * The stream ranges are sorted by <code>[stream][epoch][startOffset]</code>
      */
     private List<ObjectStreamRange> streamRanges;
 
     /**
-     * The stream objects which split from the compacted object.
+     * The stream objects which split from the WAL object.
+     * <p>
+     * The stream objects are sorted by <code>[stream][startOffset]</code>
      */
     private List<StreamObject> streamObjects;
-    private long orderId;
+
+    /**
+     * The object ids which are compacted by the WAL object.
+     */
+    private List<Long> compactedObjectIds;
 
     public long getObjectId() {
         return objectId;
@@ -49,6 +72,17 @@ public class CommitWALObjectRequest {
 
     public void setObjectSize(long objectSize) {
         this.objectSize = objectSize;
+    }
+
+    public void setCompactedObjectIds(List<Long> compactedObjectIds) {
+        this.compactedObjectIds = compactedObjectIds;
+    }
+
+    public List<Long> getCompactedObjectIds() {
+        if (compactedObjectIds == null) {
+            return Collections.emptyList();
+        }
+        return compactedObjectIds;
     }
 
     public List<ObjectStreamRange> getStreamRanges() {
@@ -98,11 +132,12 @@ public class CommitWALObjectRequest {
     @Override
     public String toString() {
         return "CommitWALObjectRequest{" +
-                "objectId=" + objectId +
-                ", objectSize=" + objectSize +
-                ", streamRanges=" + streamRanges +
-                ", streamObjects=" + streamObjects +
-                ", orderId=" + orderId +
-                '}';
+            "objectId=" + objectId +
+            ", orderId=" + orderId +
+            ", objectSize=" + objectSize +
+            ", streamRanges=" + streamRanges +
+            ", streamObjects=" + streamObjects +
+            ", compactedObjectIds=" + compactedObjectIds +
+            '}';
     }
 }
